submited one

+++ src/main/java/com/dns/crossover/ass/repo/IOrderRepo.java
@@ -11,6 +11,6 @@ import com.dns.crossover.ass.domain.Order;
  * @author dinusha
  *
  */
-public interface IOrderRepo extends CrudRepository<Order, Integer> {
+public interface IOrderRepo extends CrudRepository<Order, Long> {
 
 }
